requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
idna==3.3
model-mommy==2.0.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
